switch emoji detection to external lib gomoji
